backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.0
supabase==2.10.0
anthropic==0.40.0
gitpython==3.1.43
aiosqlite==0.20.0
